+++ src/eac3/audfrm.rs
@@ -635,6 +635,7 @@ mod tests {
             extpgmscl: None,
             paninfo: None,
             paninfo2: None,
+            premix_compression: None,
             compr: None,
             compr_ch2: None,
             dsurexmod: None,